+++ test/td_df_lib/format_test.exs
@@ -1,7 +1,6 @@
 defmodule TdDfLib.FormatTest do
   use ExUnit.Case
 
-  alias TdCache.DomainCache
   alias TdCache.Redix
   alias TdCache.SystemCache
   alias TdDfLib.Format
@@ -644,7 +643,7 @@ defmodule TdDfLib.FormatTest do
   end
 
   describe "search_values/2" do
-    setup [:create_system, :create_domain]
+    setup :create_system
 
     test "search_values/2 sets default values and removes redundant fields" do
       content = %{
@@ -726,8 +725,8 @@ defmodule TdDfLib.FormatTest do
       assert %{"system" => [_system]} = Format.search_values(content, %{content: fields})
     end
 
-    test "search_values/2 gets domain from cache and formats it", %{domain: %{id: domain_id}} do
-      content = %{"domain" => domain_id}
+    test "search_values/2 returns unchanged domain fields" do
+      content = %{"domain" => 123}
 
       fields = [
         %{
@@ -738,9 +737,9 @@ defmodule TdDfLib.FormatTest do
         }
       ]
 
-      assert %{"domain" => ^domain_id} = Format.search_values(content, %{content: fields})
+      assert %{"domain" => 123} = Format.search_values(content, %{content: fields})
 
-      content = %{"domain" => [domain_id]}
+      content = %{"domain" => [123, 456]}
 
       fields = [
         %{
@@ -751,7 +750,7 @@ defmodule TdDfLib.FormatTest do
         }
       ]
 
-      assert %{"domain" => [^domain_id]} = Format.search_values(content, %{content: fields})
+      assert %{"domain" => [123, 456]} = Format.search_values(content, %{content: fields})
     end
 
     test "search_values/2 returns nil when no template is provided" do
@@ -831,15 +830,12 @@ defmodule TdDfLib.FormatTest do
   end
 
   describe "enrich_content_values/2" do
-    setup [:create_system, :create_domain]
+    setup :create_system
 
-    test "enrich_content_values/2 gets cached values for system fields", %{
-      domain: %{id: domain_id},
-      system: system
-    } do
+    test "enrich_content_values/2 gets cached values for system fields", %{system: system} do
       content = %{
         "system" => %{"id" => system.id},
-        "domain" => domain_id,
+        "domain" => 123,
         "foo" => "bar"
       }
 
@@ -854,7 +850,7 @@ defmodule TdDfLib.FormatTest do
         }
       ]
 
-      assert %{"system" => ^system, "domain" => ^domain_id, "foo" => "bar"} =
+      assert %{"system" => ^system, "domain" => 123, "foo" => "bar"} =
                Format.enrich_content_values(content, %{content: fields})
     end
   end
@@ -868,24 +864,6 @@ defmodule TdDfLib.FormatTest do
       Redix.command(["DEL", "systems:ids_external_ids"])
     end)
 
-    {:ok, system: system}
-  end
-
-  defp create_domain(_) do
-    domain = %{
-      id: System.unique_integer([:positive]),
-      external_id: "foo",
-      name: "bar",
-      updated_at: DateTime.utc_now()
-    }
-
-    DomainCache.put(domain)
-
-    on_exit(fn ->
-      DomainCache.delete(domain.id)
-      Redix.command(["DEL", "domains:ids_to_external_ids"])
-    end)
-
-    {:ok, domain: domain}
+    [system: system]
   end
 end